home component and navbar created and working
