+++ src/components/Navbar/Navbar.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+class Navbar extends React.Component {
+  render() {
+    return (
+      <h2>Navbar Component</h2>
+    );
+  }
+}
+
+export default Navbar;
